requirements.txt
discord.py==1.7.3
steamspypi==1.1.1